refactor: put InMemorySessionStoreTest in data package
