+++ features/select-doc/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internalapi/nav/SelectDocDestinations.kt
@@ -18,4 +18,7 @@ sealed interface SelectDocDestinations : ProveYourIdentityDestinations {
 
     @Serializable
     data object ConfirmPassport : SelectDocDestinations
+
+    @Serializable
+    data object ConfirmBrp : SelectDocDestinations
 }
