+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/SelectDocNavGraphProvider.kt
@@ -9,7 +9,6 @@ import com.squareup.anvil.annotations.ContributesMultibinding
 import uk.gov.onelogin.criorchestrator.features.resume.internalapi.nav.ProveYourIdentityNavGraphProvider
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp.select.SelectBrpScreen
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.drivinglicence.select.SelectDrivingLicenceScreen
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.confirmation.ConfirmDocumentScreen
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.confirmation.ConfirmPassportScreen
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.confirmation.ConfirmPassportViewModelModule
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.passport.select.SelectPassportScreen
@@ -26,51 +25,51 @@ import javax.inject.Named
 
 @ContributesMultibinding(CriOrchestratorScope::class)
 class SelectDocNavGraphProvider
-@Inject
-constructor(
-    @Named(SelectPassportViewModelModule.FACTORY_NAME)
-    private val selectPassportViewModelFactory: ViewModelProvider.Factory,
-    @Named(SelectBrpViewModelModule.FACTORY_NAME)
-    private val selectBrpViewModelFactory: ViewModelProvider.Factory,
-    @Named(ConfirmPassportViewModelModule.FACTORY_NAME)
-    private val confirmPassportViewModelFactory: ViewModelProvider.Factory,
-    @Named(TypesOfPhotoIDViewModelModule.FACTORY_NAME)
-    private val typesOfPhotoIDViewModelFactory: ViewModelProvider.Factory,
-) : ProveYourIdentityNavGraphProvider {
-    override fun NavGraphBuilder.contributeToGraph(navController: NavController) {
-        composable<SelectDocDestinations.Passport> {
-            SelectPassportScreen(
-                navController = navController,
-                viewModel = viewModel(factory = selectPassportViewModelFactory),
-            )
-        }
+    @Inject
+    constructor(
+        @Named(SelectPassportViewModelModule.FACTORY_NAME)
+        private val selectPassportViewModelFactory: ViewModelProvider.Factory,
+        @Named(SelectBrpViewModelModule.FACTORY_NAME)
+        private val selectBrpViewModelFactory: ViewModelProvider.Factory,
+        @Named(ConfirmPassportViewModelModule.FACTORY_NAME)
+        private val confirmPassportViewModelFactory: ViewModelProvider.Factory,
+        @Named(TypesOfPhotoIDViewModelModule.FACTORY_NAME)
+        private val typesOfPhotoIDViewModelFactory: ViewModelProvider.Factory,
+    ) : ProveYourIdentityNavGraphProvider {
+        override fun NavGraphBuilder.contributeToGraph(navController: NavController) {
+            composable<SelectDocDestinations.Passport> {
+                SelectPassportScreen(
+                    navController = navController,
+                    viewModel = viewModel(factory = selectPassportViewModelFactory),
+                )
+            }
 
-        composable<SelectDocDestinations.Brp> {
-            SelectBrpScreen(
-                navController = navController,
-                viewModel = viewModel(factory = selectBrpViewModelFactory),
-            )
-        }
+            composable<SelectDocDestinations.Brp> {
+                SelectBrpScreen(
+                    navController = navController,
+                    viewModel = viewModel(factory = selectBrpViewModelFactory),
+                )
+            }
 
-        composable<SelectDocDestinations.DrivingLicence> {
-            SelectDrivingLicenceScreen()
-        }
+            composable<SelectDocDestinations.DrivingLicence> {
+                SelectDrivingLicenceScreen()
+            }
 
-        composable<SelectDocDestinations.TypesOfPhotoID> {
-            TypesOfPhotoIDScreen(
-                viewModel = viewModel(factory = typesOfPhotoIDViewModelFactory),
-            )
-        }
+            composable<SelectDocDestinations.TypesOfPhotoID> {
+                TypesOfPhotoIDScreen(
+                    viewModel = viewModel(factory = typesOfPhotoIDViewModelFactory),
+                )
+            }
 
-        composable<SelectDocDestinations.ConfirmPassport> {
-            ConfirmPassportScreen(
-                navController = navController,
-                viewModel = viewModel(factory = confirmPassportViewModelFactory),
-            )
-        }
+            composable<SelectDocDestinations.ConfirmPassport> {
+                ConfirmPassportScreen(
+                    navController = navController,
+                    viewModel = viewModel(factory = confirmPassportViewModelFactory),
+                )
+            }
 
-        composable<SelectDocDestinations.ConfirmBrp> {
-            ConfirmBrpScreen()
+            composable<SelectDocDestinations.ConfirmBrp> {
+                ConfirmBrpScreen()
+            }
         }
     }
-}
